LayoutTests/imported/w3c/web-platform-tests/fetch/http-cache/partial-expected.txt


FAIL HTTP cache stores partial content and reuses it. assert_equals: Response body expected "3198c063-b8ad-4992-8462-70bf7483cc64" but got "01234"
FAIL HTTP cache stores complete response and serves smaller ranges from it. assert_less_than: Response used expected a number less than 2 but got 2
FAIL HTTP cache stores partial response and serves smaller ranges from it. assert_less_than: Response used expected a number less than 2 but got 2
FAIL HTTP cache stores partial content and completes it. assert_equals: expected (string) "bytes=5-" but got (undefined) undefined
